+++ package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "@material-ui/core": "4.10.0",
     "@material-ui/icons": "4.9.1",
+    "axios": "^0.19.2",
     "chartist": "0.10.1",
     "classnames": "2.2.6",
     "firebase": "^7.14.6",
@@ -16,6 +17,7 @@
     "react": "16.13.1",
     "react-chartist": "0.14.3",
     "react-dom": "16.13.1",
+    "react-google-login": "^5.1.20",
     "react-google-maps": "9.4.5",
     "react-router-dom": "5.2.0",
     "react-scripts": "3.4.1",
