+++ src/firebase-config.js
@@ -1,0 +1,13 @@
+// Firebase project Configuration
+const firebaseConfig = {
+  apiKey: "AIzaSyATjcR_5F98ADo3iY5ZxgpcBSbMX4Z_M34",
+  authDomain: "csa-app-41a53.firebaseapp.com",
+  databaseURL: "https://csa-app-41a53.firebaseio.com",
+  projectId: "csa-app-41a53",
+  storageBucket: "csa-app-41a53.appspot.com",
+  messagingSenderId: "796324132621",
+  appId: "1:796324132621:web:e5a858cb09ca3eb2a1e244",
+  measurementId: "G-QT2TLBEY8V",
+};
+
+export default firebaseConfig;
